requirements.txt
django==4.2.4
djangorestframework==3.14.0
django-allauth==0.54.0
dj-rest-auth==4.0.1
django-environ==0.10.0
psycopg2-binary==2.9.7
django-model-utils==4.3.1
django-rest-knox==4.2.0

# Models
faker==19.3.0
ulid~=1.1
ulid-py==1.1.0

# QR code
Pillow==10.0.0
qrcode[PIL]==7.4.2